fix: address DVL-75 review feedback
