--- kyubic_ws/src/driver/blue_rov/dvl75_driver/dvl75_driver/__init__.py
@@ -1,1 +1,0 @@
-"""Cerulean Sonar DVL-75 ROS 2 driver package."""
