+++ kyubic_ws/src/driver/blue_rov/dvl75_driver/package.xml
@@ -8,7 +8,6 @@
   <license>Apache License 2.0</license>
 
   <buildtool_depend>ament_cmake</buildtool_depend>
-  <buildtool_depend>ament_cmake_python</buildtool_depend>
 
   <depend>blue_rov_msgs</depend>
   <depend>builtin_interfaces</depend>
@@ -16,13 +15,11 @@
   <depend>custom_socket</depend>
   <depend>driver_msgs</depend>
   <depend>geometry_msgs</depend>
-  <depend>rclpy</depend>
   <depend>rclcpp</depend>
   <depend>rclcpp_components</depend>
 
   <test_depend>ament_cmake_gtest</test_depend>
   <test_depend>ament_cmake_clang_format</test_depend>
-  <test_depend>ament_cmake_pytest</test_depend>
 
   <export>
     <build_type>ament_cmake</build_type>
