--- kyubic_ws/src/driver/blue_rov/dvl75_driver/ruff.toml
@@ -1,7 +1,0 @@
-line-length = 100
-
-[lint]
-select = ["D", "E", "F", "I"]
-
-[lint.pydocstyle]
-convention = "google"
